refactor: wind extrapolation functionality
