+++ src/geodata/model/wind/extrapolate.py
@@ -1,0 +1,213 @@
+# Copyright 2023 Michael Davidson (UCSD), Xiqiang Liu (UCSD)
+
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License as
+# published by the Free Software Foundation; either version 3 of the
+# License, or (at your option) any later version.
+
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+from pathlib import Path
+from typing import Optional
+
+import numpy as np
+import pandas as pd
+import xarray as xr
+
+from ...config import DATASET_ROOT_PATH
+from ...logging import logger
+from ._base import HEIGHTS, WindBaseModel
+
+try:
+    from numba import njit, prange
+except ImportError:
+    logger.warning("Numba not installed. Using pure Python implementation.")
+    prange = range
+
+    from ...utils import dummy_njit as njit
+
+
+@njit(parallel=True)
+def _compute_wind_speed_ext(
+    heights: np.ndarray, speeds: np.ndarray
+) -> tuple[np.ndarray, np.ndarray]:
+    """Compute wind speed from heights and speeds.
+
+    Args:
+        heights (np.ndarray): Array of heights.
+        speeds (np.ndarray): Array of speeds.
+
+    Returns:
+        tuple[np.ndarray, np.ndarray]: Tuple of coefficients (alpha, beta) and residuals.
+    """
+    # pylint: disable=not-an-iterable
+    coeffs = np.empty(shape=heights.shape[:-1] + (2,))
+    residuals = np.empty_like(speeds)
+    for time in prange(heights.shape[0]):
+        for lat in prange(heights.shape[1]):
+            for lon in prange(heights.shape[2]):
+                a = heights[time, lat, lon]
+                a = np.stack((a, np.ones_like(a)), axis=-1)
+                b = np.abs(speeds[time, lat, lon])
+
+                coeff, residual, _, _ = np.linalg.lstsq(a, b)
+
+                coeffs[time, lat, lon] = coeff
+                residuals[time, lat, lon] = residual
+    # pylint: enable=not-an-iterable
+
+    return coeffs, residuals
+
+
+class WindExtrapolationModel(WindBaseModel):
+    """Wind speed estimation based on the an extrapolation model.
+
+    Model Details:
+        The wind speed is estimated using a linear regression of the logarithm of the wind speed
+        against the logarithm of the height. The coefficients of the regression are stored in the
+        dataset as a 2D array of shape (2,). The first coefficient is the slope of the regression
+        line, and the second coefficient is the intercept. The wind speed is then estimated as
+        :math:`\\hat{v} = \\alpha \\cdot \\log(z) + \\beta`, where :math:`\\alpha` and :math:`\\beta`
+        are the coefficients, and :math:`z` is the height.
+
+    Residuals:
+        The residuals of the regression are stored in the dataset as a 2D array of shape (n,).
+        The residuals are the difference between the estimated wind speed and the actual wind speed.
+        The residuals are stored in the same order as the heights used in the regression.
+
+    Example:
+        >>> from geodata import Dataset
+        >>> from geodata.model.wind import WindExtrapolationModel
+        >>> dataset = Dataset(module="merra2", weather_data_config="surface_flux_hourly", years=slice(2010, 2010), months=slice(1,2))
+        >>> model = WindExtrapolationModel(dataset)
+        >>> model.prepare()
+        >>> model.estimate(xs=slice(1, 2), ys=slice(1, 2), years=slice(2010, 2010), months=slice(1, 2))
+    """
+
+    def _prepare_fn(
+        self,
+        ds: xr.Dataset,
+        compute_lml: bool = True,
+        half_precision: bool = True,
+        compute_residuals: bool = True,
+    ) -> xr.Dataset:
+        """Compute wind speed from for MERRA2 dataset.
+
+        Args:
+            ds (xr.Dataset): MERRA2 dataset.
+            compute_lml (bool, optional): Include speed at LML in calculation. Defaults to True.
+            half_precision (bool, optional): Use float32 precision to store coefficients and residuals. Defaults to True.
+            compute_residuals (bool, optional): Compute and store residuals against existing data points. Defaults to True.
+
+        Returns:
+            xr.Dataset: Dataset with wind speed.
+        """
+
+        disph = ds["disph"].values
+
+        variables = [f for f in HEIGHTS if f in ds and f.replace("u", "v") in ds]
+        heights = np.array([HEIGHTS[f] for f in variables]) - disph[..., np.newaxis]
+
+        if compute_lml:
+            hlml = ds["hlml"].values
+            variables.append("ulml")
+            heights = np.concatenate(
+                (heights, (hlml - disph)[..., np.newaxis]), axis=-1
+            )
+        log_heights = np.log(heights, out=-np.ones_like(heights), where=heights > 0)
+
+        speeds = []
+        for var in variables:
+            speeds.append(
+                ((ds[var] ** 2 + ds[var.replace("u", "v")] ** 2) ** 0.5).values
+            )
+        speeds = np.stack(speeds, axis=-1).astype(heights.dtype)
+
+        coeffs, residuals = _compute_wind_speed_ext(log_heights, speeds)
+
+        if half_precision:
+            coeffs = coeffs.astype("float32")
+            residuals = residuals.astype("float32")
+        ds = ds.assign_coords(coeff=["alpha", "beta"])
+        ds["coeffs"] = (("time", "lat", "lon", "coeff"), coeffs)
+
+        if compute_residuals:
+            ds = ds.assign_coords(residual=variables)
+            ds["residuals"] = (("time", "lat", "lon", "residual"), residuals)
+
+        return ds[["coeffs", "residuals"]] if compute_residuals else ds[["coeffs"]]
+
+    # pylint: disable=arguments-differ
+    def _estimate_dataset(
+        self,
+        height: int,
+        years: slice,
+        months: Optional[slice] = None,
+        xs: Optional[slice] = None,
+        ys: Optional[slice] = None,
+        use_real_data: Optional[bool] = False,
+    ) -> xr.Dataset:
+        assert height > 0, "Height must be greater than 0."
+
+        if months is None:
+            months = slice(1, 12)
+
+        raw_paths = [Path(p) for c, p in self.metadata["files"]]
+        ds_paths = []
+
+        for p in raw_paths:
+            assert (self._path / p).exists(), f"File {p} does not exist."
+
+            try:
+                year = int(p.parts[1])
+                month = int(p.parts[2])
+            except ValueError:
+                logger.warning(
+                    "Illegal file %s contained in model, final estimated result may be incomplete",
+                    p,
+                )
+
+            if year not in range(years.start, years.stop + 1) or month not in range(
+                months.start, months.stop + 1
+            ):
+                continue
+
+            orig_ds_path = DATASET_ROOT_PATH / self.module / p.relative_to(p.parts[0])
+            ds_paths += [
+                self._path / p,
+                orig_ds_path.with_name(orig_ds_path.name.replace(".params", "")),
+            ]
+
+        start_time = pd.Timestamp(year=years.start, month=months.start, day=1)
+        end_time = pd.Timestamp(
+            year=years.stop, month=months.stop + 1, day=1
+        )  # exclude last month
+
+        ds = xr.open_mfdataset(ds_paths)
+
+        if xs is None:
+            xs = ds.coords["lon"]
+        if ys is None:
+            ys = ds.coords["lat"]
+
+        ds = ds.sel(lon=xs, lat=ys, time=slice(start_time, end_time))
+
+        if height in HEIGHTS.values() and use_real_data:
+            logger.info("Using real data for estimation at height %d", height)
+            return (ds[f"u{height}m"] ** 2 + ds[f"v{height}m"] ** 2) ** 0.5
+
+        alpha = ds["coeffs"][..., 0]
+        beta = ds["coeffs"][..., 1]
+
+        return alpha * np.log((height - ds["disph"]) / np.exp(-beta / alpha))
+
+    def _estimate_cutout(self, xs: slice, ys: slice, ts: slice) -> xr.Dataset:
+        raise NotImplementedError
+
+    # pylint: enable=arguments-differ
